Fix broken syntax highlighting after grammar installation.
When the default configuration of prompting to install the grammar was
used, it caused (font locking) buffer local variables to be
initialized based on the minibuffer-mode major mode being activated.
This prevented the font lock settings for gpr-ts-mode from being
applied, immediately after the grammar was installed.  Now, read-key
is used instead of read-from-minibuffer which prevents these variables
from being initialized.