Reword commentary section and package brief description.
The commentary wording implied that the user should ensure the grammar
is installed prior to using the mode.  This is misleading as the
grammar can be installed when the mode is first activated.

Reworded the package brief description to indicate this this is a
major mode.  The original wording followed the template in the Emacs
intrinsic tree-sitter modes, but that description is unclear.  It
doesn't indicate to the user that it is a major mode, only that it
provided support for tree-sitter.