Update supporting indent variables after main indentation changes.
The main problem this solves is that supporting variables are not
reevaluated when a third party package is used to auto-detect the
indentation in a file and update the main indentation variable (i.e.,
`gpr-ts-mode-indent-offset') in the buffer.  When that happens the
remaining indentation variables should be reevaluated, as they are
often computed in terms of the main indentation variable.