Fix highlighting of package and project names in variable references.
There are ambiguities in variable references (such as Foo'Bar) where
it's ambiguous whether this is a reference to attribute Bar in package
Foo of the current project or a reference to a top-level attribute Bar
in project Foo.  Another ambiguity is Foo.X'Bar, as it's unclear if X
is a package in project Foo with attribute Bar or if Foo.X is a child
project and Bar is a top-level attribute of the child project.  This
issue applies to variable references as well the variable attribute
references demonstrated above, since variables can be declared both at
the project level as well as within a package.

These ambiguities can't be resolved by the syntax tree itself and
would require active checking of package declaration names in the file
or project references (with, extends, etc.) or maintaining a list of
known package names.

If we can't correctly identify package names in variable references,
that leaves a highlighting inconsistency where package declaration
names would be highlighted, but variable references wouldn't.  This
would mean that for consistency, all package name highlighting would
need to be removed.

None of these options are ideal.  Package name highlighting gives
visual cues to the reader, indicating if a variable attribute
reference is for a package or a project.  Since it seems useful for
the user to distinguish this information, the option to maintain a
list of known packages is implemented.  This list is a user option
which contains the currently known package names, however if custom
packages are created, this list can easily be updated.