Remove misbehaving invalid syntax tests.
It can be difficult to run tests for invalid syntax as these tests are
brittle and updates to the tree-sitter library can cause the behavior
to change, as the results are undefined.  These tests were passing
locally but fail under CI, likely due to a different tree-sitter
library being used.  Therefore, the tests are being removed.