Configure indentation strategy only for newline tests.
No other tests rely on line-based indenting, so the setting has no
effect being applied to them, and if it does have an effect, it's an
indication of an error.  Therefore, the configuration is moved such
that it's only applied to newline tests.  All other tests utilize
region indenting instead of line indenting.