Improve declaration-based indentation.
When an incomplete package was next to the "end" keyword of the
project, it may be considered a valid package and therefore include
the end of the project declaration in the re-indentation of the
package.  In order to prevent this from happening, the package name
and end name must now match before the declaration-based indentation
will be applied.

Additionally, to provide a hint to the user, the end name of a package
is only syntax highlighted when both package name and end name match.
This provides a visual clue to the user of the mismatched names.

Additional tests have been added to check for improper
declaration-based indentation as well as testing the syntax
highlighting for package declarations.