Additional indentation recovery for newline and case construction.
The newline indentation uses the indentation of the next non-comment
sibling node.  When this node was an ERROR, the catch-all indentation
rule was matched which placed the newline at column 0.  Now, ERROR
nodes are handled by the indentation error recovery logic and newline
indentation prior to an ERROR node will be chosen according to this
logic.

The "end" corresponding to a case construction was not being properly
indented to the same level as the beginning of the case construction,
within the indentation error recovery logic.  This has been corrected.

Additional tests have been added to check indentation error recovery
logic for "end" nodes of a project, package and case construction
declaration.  A test was also added to check for newline indentation
just prior to an ERROR node.