Simplify indentation rules.
Update indentation rules to be more generic, preventing the need to
call out each node individually.  This allows the rules to apply in
additional scenarios (such as for newlines).

Work around an issue in the tree-sitter library which doesn't allow
obtaining the next sibling node when the current node is a missing
node inserted by the parser.  This behavior seems like a bug, but the
workaround is to obtain all nodes for a parent and manually iterate
through them instead.

Reinstate invalid-syntax tests which were previously failing and
removed from the test suite.  The error recovery logic is much
improved since those tests were removed.